separate class for table values
