+++ src/main/java/presenter/PNLPresenter.java
@@ -1,40 +1,15 @@
 package presenter;
 
-import java.time.LocalDate;
-import java.util.HashMap;
+import calculations.PNLTableValues;
 
 public class PNLPresenter {
-    //private PositionAtTime positionAtTime
-    private LocalDate firstDate;
+    private PNLTableValues tableValues;
 
-    private LocalDate lastDate;
-    private HashMap<LocalDate, double[]> PNLTable;
-
-    public PNLPresenter()//add in PositionAtTime as parameter
-    {
-        //this.positionAtTime = positionAtTime
-        //this.firstDate = 1st transaction date from sp in the PositionAtTime
-        //this.lastDate = most recent
-        PNLTable = new HashMap<>();
+    public PNLPresenter (PNLTableValues tableValues) {
+        this.tableValues = tableValues;
     }
 
-    public void setLastDate(LocalDate lastDate) {
-        this.lastDate = lastDate;
-    }
-
-    //for each date from first to last date
-        //HashMap<String, Double> positions = store positionAtTime.getPositionAtTime in variable
-
-        //pnlTableValues = new double[3]
-        //pnlTableValues[0] = traverse through positions and calculate values(units of fx)*rate from that days close or current if today(see lines 63-69 in PositionAtTime)
-        //pnlTableValues[1] = value of "USD" from positions;
-        //pnlTableValues[2] = (pnlTableValues[0]+pnlTableValues[1]) - (previous)(pnlTableValues[0]+pnlTableValues[1]) OR 0 if first transaction
-
-        //PNLTable.put(date, pnlTableValues);
-
-
-   /* public double getOverallPnL(){
+    /* public double getOverallPnL(){
         ((lastDate)pnlTableValues[0]+pnlTableValues[1]) - (firstDate)(pnlTableValues[0]+pnlTableValues[1])
     }*/
-
 }
